Revert "부산대_5조_순수웨딩_5주차_MASTER"

+++ sunsu-wedding/src/main/java/com/kakao/sunsuwedding/portfolio/price/PriceItemJPARepository.java
@@ -16,5 +16,5 @@ public interface PriceItemJPARepository extends JpaRepository<PriceItem, Long> {
     @EntityGraph("PriceItemWithPortfolioAndPlanner")
     List<PriceItem> findAllByPortfolioId(Long id);
 
-    void deleteAllByPortfolioPlannerId(Long id);
+    void deleteAllByPortfolioPlannerId(int id);
 }
